fixed file organization
