+++ src/model/RocketShip_BaseColor.png
