--- src/modelRenderingTest/index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <script src="https://webglfundamentals.org/webgl/resources/webgl-utils.js"></script>
-        <script src="https://webglfundamentals.org/webgl/resources/m4.js"></script>
-        <title>WebGL Demo</title>
-        <script
-            src="https://cdnjs.cloudflare.com/ajax/libs/gl-matrix/2.8.1/gl-matrix-min.js"
-            integrity="sha512-zhHQR0/H5SEBL3Wn6yYSaTTZej12z0hVZKOv3TwCUXT1z5qeqGcXJLLrbERYRScEDDpYIJhPC1fk31gqR783iQ=="
-            crossorigin="anonymous"
-            defer>
-        </script>
-        <script src="webgl-demo.js" type="module"></script>
-        <link rel="stylesheet" href="./webgl.css" type="text/css" />
-        <link rel="stylesheet" href="styles.css">
-    </head>
-    <body>
-        <canvas id="gl-canvas"></canvas>
-    </body>
-</html>
