--- src/modelRenderingTest/model/RocketShip.mtl
@@ -1,8 +1,0 @@
-newmtl RocketShip_mat
-illum 4
-Kd 0.50 0.50 0.50
-Ka 0.00 0.00 0.00
-Tf 1.00 1.00 1.00
-Ni 1.00
-
-map_Kd RocketShip_BaseColor.png
