--- src/modelRenderingTest/styles.css
@@ -1,2 +1,0 @@
-body { margin: 0; }
-canvas { width: 100vw; height: 100vh; display: block; }
